requirements.txt
scikit-surprise
pandas
numpy
scikit-learn
streamlit
